+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 - Added GlobalIDInput graphene scalar.
 - Added Schema.register_mutation method for registering mutations.
 - Added Schema.graphql_mutation decorator for registering mutations.
+- Added `simple_graphql.auth` Django app with the following features:
+  - GraphQL middleware which can be used to authenticate queries based on the
+    HTTP authorization header.
+  - Token model for maintaining authorized sessions.
 
 ## 0.2.0